+++ src/main/java/testingui/diplomadoumss/org/core/FirstExample.java
@@ -9,7 +9,7 @@ import java.util.Map;
 public class FirstExample {
     public static void main(String[] args) {
 
-        WebDriver webDriver = DriverManager.getInstance().getWebDriver();
+        WebDriver webDriver = WebDriverManager.getInstance().getWebDriver();
         webDriver.get(PropertyAccesor.getInstance().getURL());
         webDriver.quit();
     }
