+++ src/main/java/testingui/diplomadoumss/org/manageloadpage/LoadPage.java
@@ -1,7 +1,7 @@
 package testingui.diplomadoumss.org.manageloadpage;
 
 import org.openqa.selenium.WebDriver;
-import testingui.diplomadoumss.org.core.DriverManager;
+import testingui.diplomadoumss.org.core.WebDriverManager;
 import testingui.diplomadoumss.org.managepage.login.Login;
 import testingui.diplomadoumss.org.utilsfiles.PropertyAccesor;
 
@@ -12,7 +12,7 @@ import testingui.diplomadoumss.org.utilsfiles.PropertyAccesor;
 public class LoadPage {
 
     public static Login loadPPHPTravels(){
-        WebDriver webDriver = DriverManager.getInstance().getWebDriver();
+        WebDriver webDriver = WebDriverManager.getInstance().getWebDriver();
         webDriver.get(PropertyAccesor.getInstance().getURL());
         return new Login();
     }
