+++ src/main/java/testingui/diplomadoumss/org/managepage/dashboard/Dashboard.java
@@ -20,7 +20,7 @@ public class Dashboard extends BasePage {
 
     public Dashboard() {
 //        isWebElementVisible(bookingLink);
-        avoidToUse(3);
+        avoidToUse(5);
     }
 
     public Dashboard clickCarsExpand() {
